+++ aiocraft/mc/proto/play/clientbound/packet_named_entity_spawn.py
@@ -2,6 +2,7 @@
 
 from typing import Tuple, List, Dict, Union, Optional
 from ....packet import Packet
+from ....definitions import *
 from ....types import *
 
 class PacketNamedEntitySpawn(Packet):
@@ -12,10 +13,10 @@ class PacketNamedEntitySpawn(Packet):
 	metadata : dict
 	pitch : int
 	playerUUID : str
-	x : Union[int,float]
-	y : Union[int,float]
+	x : Union[float,int]
+	y : Union[float,int]
 	yaw : int
-	z : Union[int,float]
+	z : Union[float,int]
 
 	def __init__(self, proto:int,
 		currentItem:int=None,
@@ -23,10 +24,10 @@ class PacketNamedEntitySpawn(Packet):
 		metadata:dict=None,
 		pitch:int=None,
 		playerUUID:str=None,
-		x:Union[int,float]=None,
-		y:Union[int,float]=None,
+		x:Union[float,int]=None,
+		y:Union[float,int]=None,
 		yaw:int=None,
-		z:Union[int,float]=None,
+		z:Union[float,int]=None,
 		**kwargs
 	):
 		super().__init__(proto,
