+++ aiocraft/mc/proto/play/clientbound/packet_spawn_entity_weather.py
@@ -2,6 +2,7 @@
 
 from typing import Tuple, List, Dict, Union, Optional
 from ....packet import Packet
+from ....definitions import *
 from ....types import *
 
 class PacketSpawnEntityWeather(Packet):
@@ -9,16 +10,16 @@ class PacketSpawnEntityWeather(Packet):
 	
 	entityId : int
 	type : int
-	x : Union[int,float]
-	y : Union[int,float]
-	z : Union[int,float]
+	x : Union[float,int]
+	y : Union[float,int]
+	z : Union[float,int]
 
 	def __init__(self, proto:int,
 		entityId:int=None,
 		type:int=None,
-		x:Union[int,float]=None,
-		y:Union[int,float]=None,
-		z:Union[int,float]=None,
+		x:Union[float,int]=None,
+		y:Union[float,int]=None,
+		z:Union[float,int]=None,
 		**kwargs
 	):
 		super().__init__(proto,
